// src/api/vcci-news/models/siteInformationBranch.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */

export interface SiteInformationBranch {
  id?: string;
  /** @nullable */
  branch_name?: string | null;
  /** @nullable */
  hotline?: string | null;
  /** @nullable */
  telephone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  fax?: string | null;
  /** @nullable */
  googlemap_link?: string | null;
  sort_order?: number;
  is_active?: boolean;
  created_at?: string;
  /** @nullable */
  created_by?: string | null;
  updated_at?: string;
  /** @nullable */
  updated_by?: string | null;
}