// src/api/vcci-news/models/siteInformationData.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */
import type { SiteInformationDataLogo } from './siteInformationDataLogo';
import type { SiteInformationSocialLink } from './siteInformationSocialLink';
import type { SiteInformationBranch } from './siteInformationBranch';

export interface SiteInformationData {
  id?: string;
  code?: string;
  /** @nullable */
  website_name?: string | null;
  /** @nullable */
  website_link?: string | null;
  /** @nullable */
  logo?: SiteInformationDataLogo;
  /** @nullable */
  telephone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  working_hours?: string | null;
  link_socials?: SiteInformationSocialLink[];
  branches?: SiteInformationBranch[];
  socials?: SiteInformationSocialLink[];
  created_at?: string;
  /** @nullable */
  created_by?: string | null;
  updated_at?: string;
  /** @nullable */
  updated_by?: string | null;
}